report sale

+++ app/Http/Controllers/Intranet/SaleController.php
@@ -8,6 +8,7 @@ use App\Models\Intranet\Sale;
 use App\Models\Intranet\Type;
 use App\Models\Intranet\Agency;
 use App\Models\Intranet\Status;
+use Barryvdh\DomPDF\Facade\Pdf;
 use App\Models\Intranet\Customer;
 use App\Models\Intranet\Employee;
 use App\Models\Intranet\SaleDate;
@@ -26,7 +27,7 @@ class SaleController extends ApiController
         $filters = $request->all();
         $salesQuery = Sale::filterSales($filters)
             ->with('status', 'salesChannel', 'type', 'agency', 'customer', 'employee')
-            ->orderBy('created_at', 'desc'); // Ordenar del más reciente al más viejo
+            ->orderBy('updated_at', 'desc'); // Ordenar del más reciente al más viejo
 
         $sales = $salesQuery->paginate(10);
         return $this->respond($sales);
@@ -83,20 +84,29 @@ class SaleController extends ApiController
         $currentInventoryId = $sale->inventory_id; // Suponiendo que tienes esta relación
         $newInventoryId = $request->input('inventory_id');
 
-        // Verificar si el inventory_id ha cambiado
-        if ($currentInventoryId != $newInventoryId) {
-            // Recuperar el inventario anterior
-            $previousInventory = Inventory::withTrashed()->find($currentInventoryId);
-
-            if ($previousInventory) {
-                $previousInventory->restore(); // Restaurar el inventario anterior
+        // Verificar si el request tiene cancel = 1
+        if ($request->input('cancel') == 1) {
+            // Restaurar el inventario actual
+            $currentInventory = Inventory::withTrashed()->find($currentInventoryId);
+            if ($currentInventory) {
+                $currentInventory->restore(); // Restaurar el inventario actual
             }
+        } else {
+            // Verificar si el inventory_id ha cambiado
+            if ($currentInventoryId != $newInventoryId) {
+                // Recuperar el inventario anterior
+                $previousInventory = Inventory::withTrashed()->find($currentInventoryId);
 
-            // Eliminar el nuevo inventario si existe
-            if ($newInventoryId) {
-                $newInventory = Inventory::find($newInventoryId);
-                if ($newInventory) {
-                    $newInventory->delete(); // Eliminar el nuevo inventario
+                if ($previousInventory) {
+                    $previousInventory->restore(); // Restaurar el inventario anterior
+                }
+
+                // Eliminar el nuevo inventario si existe
+                if ($newInventoryId) {
+                    $newInventory = Inventory::find($newInventoryId);
+                    if ($newInventory) {
+                        $newInventory->delete(); // Eliminar el nuevo inventario
+                    }
                 }
             }
         }
@@ -106,7 +116,6 @@ class SaleController extends ApiController
 
         return $this->respond($sale);
     }
-
 
     /**
      * Remove the specified resource from storage.
@@ -130,7 +139,6 @@ class SaleController extends ApiController
         return $this->respondSuccess();
     }
 
-
     public function getOptions()
     {
         $data = [
@@ -145,124 +153,6 @@ class SaleController extends ApiController
         ];
 
         return $this->respond($data);
-    }
-
-    //no se se esta usando
-    public function perAgency(Request $request)
-    {
-        // Obtener mes y año del request
-        $month = $request->input('month');
-        $year = $request->input('year');
-        $agency = Agency::find($request->input('agency_id'));
-
-        $employees = $agency->employees->load('sales', 'quote', 'targets'); // Carga los empleados con sus relaciones
-
-        // Obtener el ID del tipo "Venta" y "Prospección"
-        $ventaTypeId = Type::where('name', 'Venta')->where('type_key', 'target')->value('id');
-        $prospeccionTypeId = Type::where('name', 'Prospección')->where('type_key', 'target')->value('id');
-
-        // Verificar si se encontraron los tipos
-        if (!$ventaTypeId || !$prospeccionTypeId) {
-            return $this->respond(['message' => 'Tipo "Venta" o "Prospección" no encontrado'], 404);
-        }
-
-        foreach ($employees as $employee) {
-            // Ventas del mes y año
-            $salesForMonth = $employee->sales->filter(function ($sale) use ($month, $year) {
-                $saleDate = Carbon::parse($sale->date);
-                return $saleDate->month == $month && $saleDate->year == $year;
-            });
-
-            // Quotes del mes y año
-            $quotesForMonth = $employee->quote->filter(function ($quote) use ($month, $year) {
-                return $quote->created_at->month == $month && $quote->created_at->year == $year;
-            });
-
-            // Filtrar metas del mes y año
-            $targetsForMonth = $employee->targets->filter(function ($target) use ($month, $year, $ventaTypeId, $prospeccionTypeId) {
-                return $target->month == $month && $target->year == $year &&
-                    ($target->type_id == $ventaTypeId || $target->type_id == $prospeccionTypeId);
-            });
-
-            // Sumar las ventas
-            $totalSales = $salesForMonth->sum('amount');
-            $totalQuantitySold = $salesForMonth->count(); // Cantidad de ventas
-
-            // Sumar las quotes
-            $totalQuotes = $quotesForMonth->sum('amount');
-            $totalQuantityQuotes = $quotesForMonth->count(); // Cantidad de quotes
-
-            // Filtrar metas para ventas
-            $targetsForSales = $targetsForMonth->filter(function ($target) use ($ventaTypeId) {
-                return $target->type_id == $ventaTypeId;
-            });
-
-            // Filtrar metas para prospección
-            $targetsForProspeccion = $targetsForMonth->filter(function ($target) use ($prospeccionTypeId) {
-                return $target->type_id == $prospeccionTypeId;
-            });
-
-            // Sumar metas para ventas
-            $totalTargetsSales = $targetsForSales->sum('value');
-            $totalQuantityTargetsSales = $targetsForSales->sum('quantity');
-
-            // Sumar metas para prospección
-            $totalTargetsProspeccion = $targetsForProspeccion->sum('value');
-            $totalQuantityTargetsProspeccion = $targetsForProspeccion->sum('quantity');
-
-            // Calcular el porcentaje y diferencias para ventas
-            $percentageDifferenceSales = $totalTargetsSales > 0
-                ? number_format(($totalSales / $totalTargetsSales) * 100, 2)
-                : ($totalSales > 0 ? 100 : 0);
-            $differenceSales = $totalSales - $totalTargetsSales;
-
-            // Calcular el porcentaje y diferencias para prospecciones
-            $percentageDifferenceProspeccion = $totalTargetsProspeccion > 0
-                ? number_format(($totalQuotes / $totalTargetsProspeccion) * 100, 2)
-                : ($totalQuotes > 0 ? 100 : 0);
-            $differenceProspeccion = $totalQuotes - $totalTargetsProspeccion;
-
-            // Calcular diferencias de cantidad para ventas
-            $quantityDifferenceSales = $totalQuantitySold - $totalQuantityTargetsSales;
-            $quantityPercentageDifferenceSales = $totalQuantityTargetsSales > 0
-                ? number_format(($totalQuantitySold / $totalQuantityTargetsSales) * 100, 2)
-                : ($totalQuantitySold > 0 ? 100 : 0);
-
-            // Calcular diferencias de cantidad para prospecciones
-            $quantityDifferenceProspeccion = $totalQuantityQuotes - $totalQuantityTargetsProspeccion;
-            $quantityPercentageDifferenceProspeccion = $totalQuantityTargetsProspeccion > 0
-                ? number_format(($totalQuantityQuotes / $totalQuantityTargetsProspeccion) * 100, 2)
-                : ($totalQuantityQuotes > 0 ? 100 : 0);
-
-            // Agregar un resumen
-            $employee->sales_summary = [
-                'sales' => [
-                    'total_sales' => $totalSales,
-                    'total_targets' => $totalTargetsSales,
-                    'met_target' => $totalSales >= $totalTargetsSales,
-                    'percentage_difference' => $percentageDifferenceSales,
-                    'difference' => $differenceSales,
-                    'total_quantity_sold' => $totalQuantitySold,
-                    'total_quantity_targets' => $totalQuantityTargetsSales,
-                    'quantity_met' => $totalQuantitySold >= $totalQuantityTargetsSales,
-                    'quantity_percentage_difference' => $quantityPercentageDifferenceSales,
-                    'quantity_difference' => $quantityDifferenceSales,
-                ],
-                'quotes' => [
-                    'total_quotes' => $totalQuotes,
-                    'total_targets' => $totalTargetsProspeccion,
-                    'met_target' => $totalQuotes >= $totalTargetsProspeccion,
-                    'percentage_difference' => $percentageDifferenceProspeccion,
-                    'difference' => $differenceProspeccion,
-                    'total_quantity_quotes' => $totalQuantityQuotes,
-                    'total_quantity_targets' => $totalQuantityTargetsProspeccion,
-                    'quantity_met' => $totalQuantityQuotes >= $totalQuantityTargetsProspeccion,
-                    'quantity_percentage_difference' => $quantityPercentageDifferenceProspeccion,
-                    'quantity_difference' => $quantityDifferenceProspeccion,
-                ],
-            ];
-        }
-        return $this->respond($employees); // Responde con los empleados y su resumen
     }
 
     public function getAgency(Request $request)
@@ -297,7 +187,7 @@ class SaleController extends ApiController
             // Ventas del mes y año
             $salesForMonth = $employee->sales->filter(function ($sale) use ($month, $year) {
                 $saleDate = Carbon::parse($sale->date);
-                return $saleDate->month == $month && $saleDate->year == $year;
+                return $saleDate->month == $month && $saleDate->year == $year && $sale->cancel == 0;
             });
 
             // Quotes del mes y año
@@ -424,5 +314,48 @@ class SaleController extends ApiController
             'employees' => $employees, // Resumen de cada empleado
             'agency_summary' => $agency_summary, // Resumen acumulado de la sucursal
         ]);
+    }
+
+    public function createPDF(Request $request)
+    {
+        // Obtener mes y año del request
+        $month = $request->input('month');
+        $year = $request->input('year');
+        $agency = Agency::find($request->input('agency_id'));
+
+        $data = $this->getAgency($request);
+
+        // Define el nombre de la sucursal
+        $branchName = $agency->name;
+
+        // Convertir el número del mes a su nombre
+        $monthName = Carbon::create()->month($month)->format('F');
+
+        // Obtener los empleados como array
+        $employees = $data->original['employees']->toArray();
+
+        // Obtener los datos agencia como array
+        $agencyData = $data->original['agency_summary'];
+
+        // Generar el PDF nuevo
+        $pdf = Pdf::loadView('pdf.sale.agency', [
+            'agency' => $agencyData,
+            'employees' => $employees,
+            'branchName' => $branchName,
+            'monthName' => $monthName,
+            'year' => $year,
+        ])->setPaper('a4', 'landscape'); // Aquí defines el papel y la orientación
+
+        // Quitar para produccion, esto es para pruebas en postman
+        // return $pdf->download('document.pdf');
+
+        // Obtener el contenido del PDF como cadena binaria
+        $pdfContent = $pdf->output();
+
+        // Convertir el contenido a Base64
+        $pdfBase64 = base64_encode($pdfContent);
+
+        // Retornar el PDF en Base64
+        return $this->respond($pdfBase64);
     }
 }
